+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [5.9.0](https://github.com/GetStream/stream-chat-swiftui/releases/tag/5.9.0)
+_August 12, 2026_
+
 ### 🐞 Fixed
 - Fix empty message bubbles for audio attachments by rendering them with the file attachment UI [#1565](https://github.com/GetStream/stream-chat-swiftui/pull/1565)
 - Fix attachment preview being too big in quoted messages when its thumbnail is unavailable [#1564](https://github.com/GetStream/stream-chat-swiftui/pull/1564)